Quick commit and push.
